+++ talos-cli/src/commands/mod.rs
@@ -1,4 +1,5 @@
 pub mod echo;
+pub mod hz;
 pub mod nodes;
 pub mod parameters;
 pub mod topics;